Adding common functionality under one integration test case
