+++ phpunit.xml
@@ -4,12 +4,12 @@
             <directory>./tests/unit</directory>
         </testsuite>
         <testsuite name="integration">
-            <directory>./tests/integration</directory>
+            <directory>./tests/Integration</directory>
         </testsuite>
     </testsuites>
     <filter>
         <whitelist>
             <directory suffix=".php">lib</directory>
         </whitelist>
-</filter>
+    </filter>
 </phpunit>
