+++ tests/Integration/CreateClustersTaskTest.php
@@ -38,53 +38,7 @@ use OCA\FaceRecognition\Migration\AddDefaultFaceModel;
 
 use Test\TestCase;
 
-class CreateClustersTaskTest extends TestCase {
-	/** @var IAppContainer */
-	private $container;
-
-	/** @var FaceRecognitionContext Context */
-	private $context;
-
-	/** @var IUser User */
-	private $user;
-
-	/** @var IConfig Config */
-	private $config;
-
-	public function setUp() {
-		parent::setUp();
-		// Better safe than sorry. Warn user that database will be changed in chaotic manner:)
-		if (false === getenv('TRAVIS')) {
-			$this->fail("This test touches database. Add \"TRAVIS\" env variable if you want to run these test on your local instance.");
-		}
-
-		// Create user on which we will upload images and do testing
-		$userManager = OC::$server->getUserManager();
-		$username = 'testuser' . rand(0, PHP_INT_MAX);
-		$this->user = $userManager->createUser($username, 'password');
-		$this->loginAsUser($username);
-		// Get container to get classes using DI
-		$app = new App('facerecognition');
-		$this->container = $app->getContainer();
-
-		// Insantiate our context, that all tasks need
-		$appManager = $this->container->query('OCP\App\IAppManager');
-		$userManager = $this->container->query('OCP\IUserManager');
-		$rootFolder = $this->container->query('OCP\Files\IRootFolder');
-		$this->config = $this->container->query('OCP\IConfig');
-		$logger = $this->container->query('OCP\ILogger');
-		$this->context = new FaceRecognitionContext($appManager, $userManager, $rootFolder, $this->config);
-		$this->context->logger = new FaceRecognitionLogger($logger);
-	}
-
-	public function tearDown() {
-		$faceMgmtService = $this->container->query('OCA\FaceRecognition\FaceManagementService');
-		$faceMgmtService->resetAllForUser($this->user->getUID());
-
-		$this->user->delete();
-
-		parent::tearDown();
-	}
+class CreateClustersTaskTest extends IntegrationTestCase {
 
 	/**
 	 * Test that one face that was not in any cluster will be assigned new person
